refactor(AttendanceStatus): 패키치 이동
